+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Socialite provider `FakeProvider` useful for local environment.
+
 ## [0.23.0] - 2023-07-03
 
 ### Added